Moved device_buffer and host_buffer to ml-prims
